+++ packages/app/src/features/home/HomeScreen.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Anchor, Button, H1, Paragraph, Separator, XStack, YStack } from 'tamagui'
-import { CaretDown } from 'tamagui-phosphor-icons'
 import { LmThemeToggle } from 'app/src/components/various/LmThemeToggle'
 import { useToggleMainDrawer } from 'app/src/state/appState'
 import { LmAppDrawer } from 'app/src/components/layouts/LmAppDrawer'
 import { LmAppShell } from 'app/src/components/layouts/LmAppShell'
+import { CaretDownRegular } from '@tamagui-extras/core'
 
 export function HomeScreen() {
   return (
@@ -37,7 +37,7 @@ function SheetDemo() {
 
   return (
     <>
-      <Button size="$6" icon={<CaretDown />} circular onPress={() => toggleMainDrawer()} />
+      <Button size="$6" icon={<CaretDownRegular />} circular onPress={() => toggleMainDrawer()} />
       <LmAppDrawer />
     </>
   )