+++ packages/app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # app
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/core@1.1.1
+  - @tamagui-extras/date@1.1.1
+  - @tamagui-extras/form@1.1.1
+  - @tamagui-extras/rich-text@1.1.1
+  - @tamagui-extras/youtube@1.1.1
+
 ## 0.0.6
 
 ### Patch Changes