+++ packages/date/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tamagui-extras/date
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/core@1.1.1
+  - @tamagui-extras/form@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes